requirement.txt
PyQt6==6.5.0
pyqtgraph==0.13.3
ccxt==3.1.58
python-dotenv